Huy branch

+++ src/services/order.service.ts
@@ -168,6 +168,9 @@ class OrderService {
             quantity: number;
             price: number;
             discount?: number;
+            images: {
+                url: string;
+            };
         }[];
         payment_method: string;
     }) {
@@ -350,6 +353,7 @@ class OrderService {
 
 
         // Tạo tài khoản người dùng nếu không có
+        let isNewUser = false;
         if (!user_id) {
             if (!user_name || !email) {
                 throw new BadRequestError('User name and email are required');
@@ -380,6 +384,7 @@ class OrderService {
                         name: user_name,
                         password: randomPassword,
                     });
+                    isNewUser = true;
                 }
             }
         }
@@ -430,7 +435,7 @@ class OrderService {
         }
 
         // Xóa giỏ hàng của người dùng trong Elasticsearch và MongoDB
-        if (user_id) {
+        if (user_id && !isNewUser) {
             await elasticsearchService.deleteDocument('carts', cart[0]._id);
             await CartModel.findByIdAndDelete(cart[0]._id);
         }
